website/requirements.txt
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-wtf==1.2.2
werkzeug==3.1.3
python-docx==1.1.2
matplotlib==3.9.3
Pillow==11.1.0
selenium==4.27.1
webdriver-manager==4.0.2
numpy==2.2.1
PyMySQL==1.1.1
cryptography==42.0.5
gunicorn==22.0.0
python-dotenv==1.0.1
whitenoise==6.8.2
